GUI Layout setting1

+++ src/TVCS/Toon/ToonInfo.java
@@ -1,13 +1,11 @@
 package TVCS.Toon;
 
-import java.io.*;
-
 /**
  * Created by ina on 2017-06-06.
  */
 public class ToonInfo extends ObjectMetaInfo{
 
-    int toonId; //if -1, it means that id wasn't allocated.
+    long toonId; //if -1, it means that id wasn't allocated.
     String name;
     IDGenerator id_generator;
     IDGenerator update_id_generator;
